fake-profile-detector/requirements.txt
Flask==2.0.3
transformers==4.11.3
Pillow